requirements.txt
python-dotenv
langchain
langchain-google-genai
langchain-google-vertexai
langchain-community
langchain-core
langchain-text-splitters
fastapi
langserve
urllib3==1.26.16
uvicorn
sse_starlette